whitelist and thread support

+++ src/events/messageUpdate.js
@@ -1,14 +1,15 @@
-const { getBlacklistRegex } = require('../utils');
+const { getBlacklistRegExp, checkWhitelists } = require('../utils');
 
 module.exports = {
 	async execute(client, old_message, new_message) {
         if (new_message.author.bot) return;
 
         if (new_message.content !== old_message.content) {
-            const regexp = getBlacklistRegex();
+            const regexp = getBlacklistRegExp();
 
-            //make sure regexp isnt empty
-            if (regexp.source !== '(?:)' && new_message.content.match(regexp)) {
+            //if anyone tries to be a smartass and edit in uncensored words, just delete the message
+            //check whitelist and make sure regexp isnt empty
+            if (!checkWhitelists(new_message) && regexp.source !== '(?:)' && new_message.content.match(regexp)) {
                 new_message.delete();
             }
         }
